Fix event workflow suppression scope
